Scripting engine to renderer binding

+++ scripting/zfconf.h
@@ -36,8 +36,8 @@
  * integer type that suits your needs, or 'float' or 'double' if you need
  * floating point numbers */
 
-typedef float zf_cell;
-#define ZF_CELL_FMT "%f"
+typedef u_int32_t zf_cell;
+#define ZF_CELL_FMT "%d"
 
 
 /* The type to use for pointers and adresses. 'unsigned int' is usually a good
